eSpencoin 3

+++ eSpencoin3.py
@@ -13,29 +13,24 @@ logging.basicConfig(level=logging.INFO)
 bot = discord.Client()
 bot = commands.Bot(command_prefix="!", description="The Central eSpencoin Exchange")
 
-
-starting_amount = 1000
-balance_prefix = "balance "
-
-
 @bot.event
 async def on_ready():
     logging.info("Starting Up")
     for guild in bot.guilds:
-        await give_starting_role_to_all_new_members(guild)
+        await assign_starting_role_to_all_new_members(guild)
         await delete_all_orphans(guild)
     logging.info("Startup Complete!")
 
 
-async def give_starting_role_to_all_new_members(guild):
+async def assign_starting_role_to_all_new_members(guild):
     logging.info(f"Assiging Roles To Server: {guild.name}")
     starting_role = await make_or_find_role(guild, starting_amount)
     for member in guild.members:
         if is_without_role(member):
-            await give_starting_role(member, starting_role)
+            await assign_starting_role(member, starting_role)
 
 
-async def give_starting_role(member, starting_role):
+async def assign_starting_role(member, starting_role):
     await member.add_roles(starting_role)
     logging.info(f"Gave starting role to: {member.name}")
 
@@ -70,7 +65,7 @@ async def delete_if_orphan(role):
 async def on_member_join(member):
     logging.info(f"Giving Role To New Member: {member.name}")
     starting_role = await make_or_find_role(member.guild, starting_amount)
-    await give_starting_role(member, starting_role)
+    await assign_starting_role(member, starting_role)
 
 
 def is_balance_role(role):
@@ -102,23 +97,25 @@ def get_balance_role(member):
 
 async def set_balance(member, amount):
     role = get_balance_role(member)
+    if role == None:
+        return False
     await member.remove_roles(role)
-    new_role = make_or_find_role(member, amount)
+    new_role = make_or_find_role(member.guild, amount)
     await member.add_roles(new_role)
     await delete_if_orphan(role)
+    return True
 
 
 def is_valid_amount(amount):
     if (
-        isnan(sreal(new_amount))
-        or isinf(sreal(new_amount))
-        or isnan(simag(new_amount))
-        or isinf(simag(new_amount))
+        isnan(sreal(amount))
+        or isinf(sreal(amount))
+        or isnan(simag(amount))
+        or isinf(simag(amount))
     ):
         return False
     else:
         return True
-
 
 def add_amounts(amount1, amount2):
     sum_amount = str(sympify(amount1) + sympify(amount2))
@@ -127,50 +124,71 @@ def add_amounts(amount1, amount2):
     else:
         return None
 
+def subtract_amounts(amount1, amount2):
+    diff_amount = str(sympify(amount1) - sympify(amount2))
+    if is_valid_amount(diff_amount):
+        return diff_amount
+    else:
+        return None
 
-async def add_balance(member, amount):
-    balance = get_balance(member)  # TODO: IF NONE RETURN FALSE
-    new_amount = add_amounts(balance, amount)  # TODO: IF NONE RETURN FALSE
-    set_balance(member, new_amount)
-    logging.info(
-        f"Added: {amount} to: {member.name} With a Final Balance of: {new_amount}"
-    )
-    return True
+def add_balance(member, amount):
+    balance = get_balance(member)
+    if balance == None:
+        return None
+    else:
+        return add_amounts(balance, amount)
+
+def subtract_balance(member, amount):
+    balance = get_balance(member)
+    if balance == None:
+        return None
+    else:
+        return subtract_amounts(balance, amount)
+
+def is_valid_target(to_member, from_member):
+    if to_member.id == bot.user.id or to_member.id == from_member.id:
+        return False
+    else:
+        return True
+
+def is_valid_give_amount_for_member(from_member, amount):
+    if sreal(amount) < 0:
+        return False
+    elif sreal(amount) == 0:
+        return simag(amount) != 0
+    else: # sreal(amount) > 0
+        balance = get_balance(from_member)
+        if balance == None:
+            return False
+        else:
+            return sreal(balance) >= sreal(amount)
 
 
 async def giveToPlayer(to_member, from_member, amount):
-    amount = sympify(amount)  # WHY SIMPIFY HERE? UNNEEDED?
-    balance = get_balance(author)
-    try:
-        if member.id != bot.user.id and (  # TODO: IS VALID GIVE TARGET
-            sreal(amount) == 0  # TODO: IS VALID GIVE AMOUNT
-            or (sreal(balance) >= sreal(amount) and sreal(amount) >= 0)
-        ):
-            add1ok = await add_balance(member, amount)  #
-            add2ok = await add_balance(
-                author, -amount
-            )  # TODO: SUBTRACT BALANCE? IS STRNG HERE...
-            if add1ok and add2ok:
-                print("GIVE")
-                await bot.add_reaction(message, "👌")
-            else:
-                print("NO GIVE")
-                await bot.add_reaction(message, "🙅")
-                await bot.add_reaction(message, "👎")
-        else:
-            print("NO GIVE")
-            await bot.add_reaction(message, "🙅")
-            await bot.add_reaction(message, "🚫")
-    except:
-        print(sys.exc_info())
-        await bot.add_reaction(message, "🙅")
-        await bot.add_reaction(message, "❓")
-    yield
 
+    if not is_valid_target(to_member, from_member):
+        return "INVALID_TARGET" #TODO: ENUM
+    
+    if not is_valid_give_amount_for_member(from_member, amount):
+        return "INVALID_GIVE_AMOUNT"
 
-def react_to_give(message, ENUM):
+    #TODO: GETS BALANCE TWICE!!
+
+    to_balance = await add_balance(to_member, amount)
+    from_balance = await subtract_balance(from_member, amount)
+
+    if to_balance == None or from_balance == None:
+        return "INVALID_BALANCE"
+
+    await set_balance(to_member, to_balance)
+    await set_balance(to_member)
+
+async def react_to_give(message, ENUM):
     pass
 
+async def give_and_react(message, to_member, from_member, amount):
+    status = await giveToPlayer(to_member, from_member, amount)
+    await react_to_give(message, status):
 
 def scom(expr):
     return complex(sympify(expr).evalf())
@@ -204,8 +222,7 @@ async def on_reaction_add(reaction, user):
 
 @bot.command(pass_context=True, description="Gives eSpencoin to another member")
 async def give(ctx, member: discord.Member, amount: str):
-    status = await giveToPlayer(member, ctx.message.author, amount)
-    await react_to_give(message, status)
+    await give_and_react(ctx.message, member, ctx.message.author, amount)
 
 
 @bot.command(pass_context=True, description="Gives eSpencoin to everyone in a role")
@@ -214,10 +231,10 @@ async def giveall(ctx, role: discord.Role, amount: str):
         if role in member.roles:
             author = ctx.message.author
             if member.id != bot.user.id and member.id != author.id:  # IS_VALID_TARGET
-                await giveToPlayer(ctx.message, member, ctx.message.author, amount)
+                await give_and_react(ctx.message, member, author, amount)
 
 
 bot.run("NTMzNDAzOTI2MDQxNDYwNzM2.DxqjdQ.YjWbPCT8Y8hJ36jAO6Xrbyh_Vg8")
 
-# TODO: REORG WHOLE FILE
-
+# TODO: REORG WHOLE FILE (classes?)
+# TODO: CONSISTENT LOGGING
